+++ config/test/bootstrap_db_tests.php
@@ -14,6 +14,7 @@ if (! file_exists('.env')) {
 
 /** @var ContainerInterface $container */
 $container = require __DIR__ . '/../container.php';
+$config = $container->get('config');
 
-$container->get(TestHelper::class)->createTestDb();
+$container->get(TestHelper::class)->createTestDb($config['entity_manager']['connection']['path']);
 DbTest\DatabaseTestCase::setEntityManager($container->get('em'));