+++ module/Core/config/dependencies.config.php
@@ -3,7 +3,6 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Core;
 
-use Cocur\Slugify\Slugify;
 use Doctrine\Common\Cache\Cache;
 use Shlinkio\Shlink\Common\Service\PreviewGenerator;
 use Shlinkio\Shlink\Core\Response\NotFoundHandler;
@@ -46,7 +45,7 @@ return [
         Options\NotFoundShortUrlOptions::class => ['config.url_shortener.not_found_short_url'],
         Options\UrlShortenerOptions::class => ['config.url_shortener'],
 
-        Service\UrlShortener::class => ['httpClient', 'em', Options\UrlShortenerOptions::class, Slugify::class],
+        Service\UrlShortener::class => ['httpClient', 'em', Options\UrlShortenerOptions::class],
         Service\VisitsTracker::class => ['em'],
         Service\ShortUrlService::class => ['em'],
         Service\VisitService::class => ['em'],