+++ module/Core/src/Service/UrlShortener.php
@@ -3,11 +3,11 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Core\Service;
 
-use Cake\Chronos\Chronos;
-use Cocur\Slugify\SlugifyInterface;
 use Doctrine\ORM\EntityManagerInterface;
+use Fig\Http\Message\RequestMethodInterface;
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\GuzzleException;
+use GuzzleHttp\RequestOptions;
 use Psr\Http\Message\UriInterface;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\Exception\EntityDoesNotExistException;
@@ -20,8 +20,12 @@ use Shlinkio\Shlink\Core\Options\UrlShortenerOptions;
 use Shlinkio\Shlink\Core\Repository\ShortUrlRepository;
 use Shlinkio\Shlink\Core\Util\TagManagerTrait;
 use Throwable;
+use function array_reduce;
+use function count;
 use function floor;
 use function fmod;
+use function Functional\contains;
+use function Functional\invoke;
 use function preg_match;
 use function strlen;
 
@@ -29,67 +33,60 @@ class UrlShortener implements UrlShortenerInterface
 {
     use TagManagerTrait;
 
-    /** @deprecated */
-    public const DEFAULT_CHARS = UrlShortenerOptions::DEFAULT_CHARS;
     private const ID_INCREMENT = 200000;
 
     /** @var ClientInterface */
     private $httpClient;
     /** @var EntityManagerInterface */
     private $em;
-    /** @var SlugifyInterface */
-    private $slugger;
     /** @var UrlShortenerOptions */
     private $options;
 
-    public function __construct(
-        ClientInterface $httpClient,
-        EntityManagerInterface $em,
-        UrlShortenerOptions $options,
-        SlugifyInterface $slugger
-    ) {
+    public function __construct(ClientInterface $httpClient, EntityManagerInterface $em, UrlShortenerOptions $options)
+    {
         $this->httpClient = $httpClient;
         $this->em = $em;
         $this->options = $options;
-        $this->slugger = $slugger;
     }
 
     /**
+     * @param string[] $tags
      * @throws NonUniqueSlugException
      * @throws InvalidUrlException
      * @throws RuntimeException
      */
-    public function urlToShortCode(
-        UriInterface $url,
-        array $tags = [],
-        ?Chronos $validSince = null,
-        ?Chronos $validUntil = null,
-        ?string $customSlug = null,
-        ?int $maxVisits = null
-    ): ShortUrl {
+    public function urlToShortCode(UriInterface $url, array $tags, ShortUrlMeta $meta): ShortUrl
+    {
+        $url = (string) $url;
+
+        // First, check if a short URL exists for all provided params
+        $existingShortUrl = $this->findExistingShortUrlIfExists($url, $tags, $meta);
+        if ($existingShortUrl !== null) {
+            return $existingShortUrl;
+        }
+
         // If the URL validation is enabled, check that the URL actually exists
         if ($this->options->isUrlValidationEnabled()) {
             $this->checkUrlExists($url);
         }
-        $customSlug = $this->processCustomSlug($customSlug);
+        $this->verifyCustomSlug($meta);
 
         // Transactionally insert the short url, then generate the short code and finally update the short code
         try {
             $this->em->beginTransaction();
 
             // First, create the short URL with an empty short code
-            $shortUrl = new ShortUrl(
-                (string) $url,
-                ShortUrlMeta::createFromParams($validSince, $validUntil, null, $maxVisits)
-            );
+            $shortUrl = new ShortUrl($url, $meta);
             $this->em->persist($shortUrl);
             $this->em->flush();
 
-            // Generate the short code and persist it
-            // TODO Somehow provide the logic to calculate the shortCode to avoid the need of a setter
-            $shortCode = $customSlug ?? $this->convertAutoincrementIdToShortCode((float) $shortUrl->getId());
-            $shortUrl->setShortCode($shortCode)
-                     ->setTags($this->tagNamesToEntities($this->em, $tags));
+            // Generate the short code and persist it if no custom slug was provided
+            if (! $meta->hasCustomSlug()) {
+                // TODO Somehow provide the logic to calculate the shortCode to avoid the need of a setter
+                $shortCode = $this->convertAutoincrementIdToShortCode((float) $shortUrl->getId());
+                $shortUrl->setShortCode($shortCode);
+            }
+            $shortUrl->setTags($this->tagNamesToEntities($this->em, $tags));
             $this->em->flush();
 
             $this->em->commit();
@@ -104,14 +101,68 @@ class UrlShortener implements UrlShortenerInterface
         }
     }
 
-    private function checkUrlExists(UriInterface $url): void
+    private function findExistingShortUrlIfExists(string $url, array $tags, ShortUrlMeta $meta): ?ShortUrl
+    {
+        if (! $meta->findIfExists()) {
+            return null;
+        }
+
+        $criteria = ['longUrl' => $url];
+        if ($meta->hasCustomSlug()) {
+            $criteria['shortCode'] = $meta->getCustomSlug();
+        }
+        /** @var ShortUrl|null $shortUrl */
+        $shortUrl = $this->em->getRepository(ShortUrl::class)->findOneBy($criteria);
+        if ($shortUrl === null) {
+            return null;
+        }
+
+        if ($meta->hasMaxVisits() && $meta->getMaxVisits() !== $shortUrl->getMaxVisits()) {
+            return null;
+        }
+        if ($meta->hasValidSince() && ! $meta->getValidSince()->eq($shortUrl->getValidSince())) {
+            return null;
+        }
+        if ($meta->hasValidUntil() && ! $meta->getValidUntil()->eq($shortUrl->getValidUntil())) {
+            return null;
+        }
+
+        $shortUrlTags = invoke($shortUrl->getTags(), '__toString');
+        $hasAllTags = count($shortUrlTags) === count($tags) && array_reduce(
+            $tags,
+            function (bool $hasAllTags, string $tag) use ($shortUrlTags) {
+                return $hasAllTags && contains($shortUrlTags, $tag);
+            },
+            true
+        );
+
+        return $hasAllTags ? $shortUrl : null;
+    }
+
+    private function checkUrlExists(string $url): void
     {
         try {
-            $this->httpClient->request('GET', $url, ['allow_redirects' => [
-                'max' => 15,
-            ]]);
+            $this->httpClient->request(RequestMethodInterface::METHOD_GET, $url, [
+                RequestOptions::ALLOW_REDIRECTS => ['max' => 15],
+            ]);
         } catch (GuzzleException $e) {
             throw InvalidUrlException::fromUrl($url, $e);
+        }
+    }
+
+    private function verifyCustomSlug(ShortUrlMeta $meta): void
+    {
+        if (! $meta->hasCustomSlug()) {
+            return;
+        }
+
+        $customSlug = $meta->getCustomSlug();
+
+        /** @var ShortUrlRepository $repo */
+        $repo = $this->em->getRepository(ShortUrl::class);
+        $shortUrlsCount = $repo->count(['shortCode' => $customSlug]);
+        if ($shortUrlsCount > 0) {
+            throw NonUniqueSlugException::fromSlug($customSlug);
         }
     }
 
@@ -132,25 +183,7 @@ class UrlShortener implements UrlShortenerInterface
         return $chars[(int) $id] . $code;
     }
 
-    private function processCustomSlug(?string $customSlug): ?string
-    {
-        if ($customSlug === null) {
-            return null;
-        }
-
-        // If a custom slug was provided, make sure it's unique
-        $customSlug = $this->slugger->slugify($customSlug);
-        $shortUrl = $this->em->getRepository(ShortUrl::class)->findOneBy(['shortCode' => $customSlug]);
-        if ($shortUrl !== null) {
-            throw NonUniqueSlugException::fromSlug($customSlug);
-        }
-
-        return $customSlug;
-    }
-
     /**
-     * Tries to find the mapped URL for provided short code. Returns null if not found
-     *
      * @throws InvalidShortCodeException
      * @throws EntityDoesNotExistException
      */