+++ module/Core/src/Service/UrlShortenerInterface.php
@@ -3,7 +3,6 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Core\Service;
 
-use Cake\Chronos\Chronos;
 use Psr\Http\Message\UriInterface;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\Exception\EntityDoesNotExistException;
@@ -11,26 +10,19 @@ use Shlinkio\Shlink\Core\Exception\InvalidShortCodeException;
 use Shlinkio\Shlink\Core\Exception\InvalidUrlException;
 use Shlinkio\Shlink\Core\Exception\NonUniqueSlugException;
 use Shlinkio\Shlink\Core\Exception\RuntimeException;
+use Shlinkio\Shlink\Core\Model\ShortUrlMeta;
 
 interface UrlShortenerInterface
 {
     /**
+     * @param string[] $tags
      * @throws NonUniqueSlugException
      * @throws InvalidUrlException
      * @throws RuntimeException
      */
-    public function urlToShortCode(
-        UriInterface $url,
-        array $tags = [],
-        ?Chronos $validSince = null,
-        ?Chronos $validUntil = null,
-        ?string $customSlug = null,
-        ?int $maxVisits = null
-    ): ShortUrl;
+    public function urlToShortCode(UriInterface $url, array $tags, ShortUrlMeta $meta): ShortUrl;
 
     /**
-     * Tries to find the mapped URL for provided short code. Returns null if not found
-     *
      * @throws InvalidShortCodeException
      * @throws EntityDoesNotExistException
      */